refactor: split EpisodeDetailScreen into focused components
Extract TimelineView, IntensitySparklineView, and EpisodeActionButtons
into separate files. EpisodeDetailScreen remains the coordinator.
